pyproject.toml
[tool.poetry]
name = "scrape-links"
version = "0.1.0"
description = ""
authors = ["Sammy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
scrapy = "^2.11.2"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
